test/reftest/vendor/webkit/fast/dom/Window/replaceable.html

<!doctype html>
<script src="../../../resources/wpt-test-pre.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
<script>

description('Tests that Replaceable attributes are writable');

Object.defineProperty(window, 'testTC', {enumerable: false,configurable: false,writable: true,value: 0});
//shouldBeTrue('Object.getOwnPropertyDescriptor(window, "innerHeight").hasOwnProperty("get")');

//window.innerHeight = 42;
//shouldBe('window.innerHeight', '42');
window.testTC = 42
shouldBe('window.testTC', '42');
</script>
<script src="../../../resources/js-test-post.js"></script>
<script src="../../../resources/wpt-test-post.js"></script>
